tests/test_main_six.py
# import os
# from pathlib import Path
# import pytest
# from main_six import FileOperations
#
# @pytest.fixture
# def test_file_path(tmp_path):
#     """Fixture to create a temporary test file."""
#     file_path = tmp_path / "test_file.txt"
#     yield file_path
#     file_path.unlink()
#
# def test_get_path(test_file_path):
#     """Test get_path method."""
#     path = FileOperations.get_path(test_file_path)
#     assert isinstance(path, Path)
#     assert path == test_file_path
#
# def test_parse_path(test_file_path):
#     """Test parse_path method."""
#     FileOperations.get_path(test_file_path)
#     directory, file_name = FileOperations.parse_path()
#     assert isinstance(directory, Path)
#     assert isinstance(file_name, str)
#     assert directory == test_file_path.parent
#     assert file_name == test_file_path.name
#
# def test_create_resources(test_file_path):
#     """Test create_resources method."""
#     FileOperations.get_path(test_file_path)
#     directory, file_name = FileOperations.parse_path()
#     FileOperations.create_resources(directory, file_name)
#     assert os.path.exists(test_file_path)
#
# def test_cleanup(test_file_path):
#     """Test cleanup method."""
#     FileOperations.get_path(test_file_path)
#     directory, file_name = FileOperations.parse_path()
#     FileOperations.create_resources(directory, file_name)
#     FileOperations.cleanup(directory, file_name)
#     assert not os.path.exists(test_file_path)
#
# def test_execute_all(test_file_path):
#     """Test execute_all method."""
#     result = FileOperations.execute_all(test_file_path)
#     assert "Operations executed successfully" in result
#     assert os.path.exists(test_file_path) == False
